Refactor to use dwolla/jenkins-agent-core:debian upstream; update nvm to 0.34.0

+++ install.sh
@@ -4,8 +4,13 @@ set +x
 echo running \`. ${NVM_DIR}/nvm.sh --no-use\` > /dev/stderr
 . ${NVM_DIR}/nvm.sh --no-use
 
-echo running \`nvm install\` > /dev/stderr
-nvm install
+if [[ -f ".nvmrc" ]]; then
+  echo running \`nvm install\` > /dev/stderr
+  nvm install
+else
+  echo running \`nvm install --lts\` > /dev/stderr
+  nvm install --lts
+fi
 set -x
 
 npm install
